CRUDDjango/requirements.txt
Pillow==5.0.0
django==2.2.12
django-ckeditor==5.4.0